refactor: extract shared polling timer implementation
Moves the polling timer logic to shared/polling-timer.js and updates
both main/ (CommonJS) and src/ (ES module) to re-export from it.

Closes #111
