+++ shared/polling-timer.js
@@ -1,0 +1,37 @@
+/**
+ * Reusable polling timer with start/stop lifecycle.
+ * Shared by both main process (CommonJS) and renderer process (ES module via esbuild).
+ *
+ * Usage:
+ *   const { PollingTimer } = require('../shared/polling-timer');
+ */
+class PollingTimer {
+  /**
+   * @param {number} intervalMs  - Polling interval in milliseconds
+   * @param {Function} callback  - Called on each tick and immediately on start
+   */
+  constructor(intervalMs, callback) {
+    this._intervalMs = intervalMs;
+    this._callback = callback;
+    this._timer = null;
+  }
+
+  start() {
+    if (this._timer) return;
+    this._timer = setInterval(() => this._callback(), this._intervalMs);
+    this._callback();
+  }
+
+  stop() {
+    if (this._timer) {
+      clearInterval(this._timer);
+      this._timer = null;
+    }
+  }
+
+  get running() {
+    return this._timer !== null;
+  }
+}
+
+module.exports = { PollingTimer };
